requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=10.0.0
tqdm>=4.65.0
einops>=0.6.0
kornia>=0.7.0
timm>=0.9.0
